feat: Parameterized DOCKER_METADATA_PR_HEAD_SHA to allow for PR-HEAD tagging (#41)
* Parameterized DOCKER_METADATA_PR_HEAD_SHA to allow for PR-HEAD tagging

By default, Github Action uses merge commits during PR events for the
(referenced via the `github.sha` variable). Some users find this
challenging, since the merge commit contains code from the target branch,
which is troublesome for long-running branches, merge-conflicts, and
a hotfix workflow for deploying one-off containers.

This allows users to configure docker metadata-action to compute the SHA
tags appropriately. Unfortunately the metadata-action does not compute
based on a checkout action of the code, so we have to tell it to use
the correct event SHA explicitly with this environment variable.

See https://github.com/docker/metadata-action#environment-variables

* Fix the datatype for the linter

Co-authored-by: Dan Miller <miller0daniel@gmail.com>

---------

Co-authored-by: Nathanael Beisiegel <pknb.dev@gmail.com>
Co-authored-by: Dan Miller <miller0daniel@gmail.com>